Make Entity class use Model instead of Mesh

+++ src/def/entity/ModelAnimation.h
@@ -15,6 +15,6 @@ public:
 
     std::string name = "";
     std::vector<AnimChannel> channels {};
-    double duration = 0;
+    uint duration = 0;
     double ticksPerSecond = 0;
 };
